Add tests and fixes to model. Add updated example setup and single notebooks

+++ Starfish/grid_tools/base_interfaces.py
@@ -296,7 +296,7 @@ class HDF5Creator:
 
         # the raw wl from the spectral library
         self.wl_native = self.grid_interface.wl  # raw grid
-        self.dv_native = calculate_dv(self.wl_native)
+        self.dv_native = calculate_dv(torch.DoubleTensor(self.wl_native))
 
         self.hdf5 = h5py.File(self.filename, "w")
         self.hdf5.attrs["grid_name"] = grid_interface.name
